requirements.txt
django-crispy-forms
pillow
django-rest-framework
django-filter
psycopg2-binary